Inputs type (#27)
* fix : inputs type set to int

* doc : CHANGELOG updated

+++ mytimer/__main__.py
@@ -19,9 +19,9 @@ def main():
         "face": 1,
         "message": ""}
     parser = argparse.ArgumentParser()
-    parser.add_argument('--minute', help='minute', type=float)
-    parser.add_argument('--second', help='second', type=float)
-    parser.add_argument('--hour', help='hour', type=float)
+    parser.add_argument('--minute', help='minute', type=int)
+    parser.add_argument('--second', help='second', type=int)
+    parser.add_argument('--hour', help='hour', type=int)
     parser.add_argument('--message', help='message', type=str)
     parser.add_argument(
         '--face',
